move enabling addressing certification in an administration section.
